Fix sticky cache
In case of error, the deduplication entry is not
removed and therefore stale mount points are hanging
around. Therefore, remove deduplicaiton entry in case
of error such that stale mounts are queued again.

Also, if not found entry in the cache, cache pod does not
provide the resolution. Therefore remove entry from deduplication.