refactor Dockerfile to use scratch as image
This refactors the build and the Dockerfile to achieve several goals:
- avoid using a base image with various packages
- avoid vulnerabilities in the base image completely
- reduce the size of the image from 138 MB down to 28 MB

The problem with the build based on the alpine image is that there will
always be some vulnerabilities in the included packages, regardless
of how minimal the images are.

This change refactors the Dockerfile and the build in such a manner that
there's no need to have a base image anymore. This avoids the issues
with vulnerabilities in the packages bundled in the base image. These
vulnerabilities are flagged by security scanners and are not acceptable.

The vulnerabilities in the Go standard library and in the used Go
packages still need to be handled accordingly even with these changes.

The build script receives these other changes:
- use bash's unofficial strict mode for better handling
- make some adjustments to the script based on Google's bash
  guide
- make some changes to how the Docker images are built, tagged and
  pushed to be more idiomatic
- address some issues with the stripping of the debugging symbols
- make some other adjustments to make the script smaller and cleaner
- remove the die helper function since it's not required anymore