update Go module dependencies to fix CVEs
This change updates the Kubernetes Go client to v0.30.14 to address
CVEs. It also removes the replace directives which were mixing the
versions of Kubernetes packages. The versions should be in sync to
avoid issues.

The chosen version of the Kubernetes Go client is recent enough while
avoiding the use of the latest version. The goal is to make it recent
enough to avoid CVEs and to avoid compatibility issues with some older
versions.

The previosly used versions of the Kubernetes Go client are for
Kubernetes versions which are no longer supported.